do not allow normal users to do update,delete,create

+++ src/app/shared/services/discipline.service.ts
@@ -1,11 +1,12 @@
-import {Injectable} from '@angular/core';
-import {HttpClient, HttpErrorResponse} from '@angular/common/http';
-import {catchError, map} from 'rxjs/operators';
-import {Observable, throwError} from 'rxjs';
-import {Discipline} from '../models/Discipline';
-import {StudyCourse} from '../models/StudyCourse';
-import {PageableResponse} from '../models/PageableResponse';
-import {ErrorService} from './error.service';
+import { Injectable } from '@angular/core';
+import { HttpClient, HttpErrorResponse } from '@angular/common/http';
+import { catchError, map } from 'rxjs/operators';
+import { Observable, throwError } from 'rxjs';
+import { Discipline } from '../models/Discipline';
+import { StudyCourse } from '../models/StudyCourse';
+import { PageableResponse } from '../models/PageableResponse';
+import { ErrorService } from './error.service';
+import { AuthService } from './auth.service';
 
 @Injectable({
   providedIn: 'root',
@@ -13,7 +14,7 @@ import {ErrorService} from './error.service';
 export class DisciplineService {
   SERVER_URL = 'http://localhost:9000/disciplines';
 
-  constructor(private http: HttpClient, private errorService: ErrorService) {
+  constructor(private http: HttpClient, private errorService: ErrorService, private auth: AuthService) {
   }
 
   private getDisciplines$(): Observable<Discipline[]> {
@@ -23,7 +24,7 @@ export class DisciplineService {
           return responseData;
         }),
         catchError((errorResponse) => {
-          this.errorService.errorSnackbar(errorResponse);
+          this.errorService.errorHTTPSnackbar(errorResponse);
           return throwError(errorResponse);
         })
       );
@@ -34,7 +35,7 @@ export class DisciplineService {
     this.getDisciplines$()
       .subscribe((response) => {
         response.forEach((d) => disciplines.push(d));
-      }, (error => this.errorService.errorSnackbar(error)));
+      }, (error => this.errorService.errorHTTPSnackbar(error)));
     return disciplines;
   }
 
@@ -46,7 +47,7 @@ export class DisciplineService {
           return responseData.content;
         }),
         catchError((errorResponse) => {
-          this.errorService.errorSnackbar(errorResponse);
+          this.errorService.errorHTTPSnackbar(errorResponse);
           return throwError(errorResponse);
         })
       );
@@ -57,32 +58,44 @@ export class DisciplineService {
     this.getStudyCoursesByDisciplineID$(disciplineId)
       .subscribe((response) => {
         response.forEach((s) => studyCourses.push(s));
-      }, (error => this.errorService.errorSnackbar(error)));
+      }, (error => this.errorService.errorHTTPSnackbar(error)));
     return studyCourses;
   }
 
   createDiscipline(data: object): Observable<any> {
+    if (!this.auth.isModOrAdmin) {
+      this.errorService.errorUnauthorized();
+      return new Observable<any>();
+    }
     return this.http.post(this.SERVER_URL, data).pipe(
       catchError((errorResponse: HttpErrorResponse) => {
-        this.errorService.errorSnackbar(errorResponse);
+        this.errorService.errorHTTPSnackbar(errorResponse);
         return new Observable();
       })
     );
   }
 
   deleteDiscipline(id: number): Observable<any> {
+    if (!this.auth.isModOrAdmin) {
+      this.errorService.errorUnauthorized();
+      return new Observable<any>();
+    }
     return this.http.delete(`${this.SERVER_URL}/${id}`).pipe(
       catchError((errorResponse: HttpErrorResponse) => {
-        this.errorService.errorSnackbar(errorResponse);
+        this.errorService.errorHTTPSnackbar(errorResponse);
         return new Observable();
       })
     );
   }
 
   updateDisciplineByID(id: number, data: { name: any }): Observable<any> {
+    if (!this.auth.isModOrAdmin) {
+      this.errorService.errorUnauthorized();
+      return new Observable<any>();
+    }
     return this.http.put(`${this.SERVER_URL}/${id}`, data).pipe(
       catchError((errorResponse: HttpErrorResponse) => {
-        this.errorService.errorSnackbar(errorResponse);
+        this.errorService.errorHTTPSnackbar(errorResponse);
         return new Observable();
       })
     );
